+++ bpdm-common/src/main/kotlin/org/eclipse/tractusx/bpdm/common/dto/IPageDto.kt
@@ -17,11 +17,24 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package org.eclipse.tractusx.orchestrator.api.client
+package org.eclipse.tractusx.bpdm.common.dto
 
-interface OrchestrationApiClient {
+import io.swagger.v3.oas.annotations.media.Schema
 
-    val goldenRecordTasks: GoldenRecordTaskApiClient
+interface IPageDto<T> {
 
-    val finishedTaskEvents: FinishedTaskEventApiClient
+    @get:Schema(description = "Total number of all results in all pages")
+    val totalElements: Long
+
+    @get:Schema(description = "Total number pages")
+    val totalPages: Int
+
+    @get:Schema(description = "Current page number")
+    val page: Int
+
+    @get:Schema(description = "Number of results in the page")
+    val contentSize: Int
+
+    @get:Schema(description = "Collection of results in the page")
+    val content: Collection<T>
 }