+++ bpdm-orchestrator-api/src/main/kotlin/org/eclipse/tractusx/orchestrator/api/client/FinishedTaskEventApiClient.kt
@@ -19,9 +19,18 @@
 
 package org.eclipse.tractusx.orchestrator.api.client
 
-interface OrchestrationApiClient {
+import org.eclipse.tractusx.bpdm.common.dto.PaginationRequest
+import org.eclipse.tractusx.orchestrator.api.FinishedTaskEventApi
+import org.eclipse.tractusx.orchestrator.api.model.FinishedTaskEventsResponse
+import org.springdoc.core.annotations.ParameterObject
+import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.service.annotation.GetExchange
+import org.springframework.web.service.annotation.HttpExchange
+import java.time.Instant
 
-    val goldenRecordTasks: GoldenRecordTaskApiClient
+@HttpExchange(FinishedTaskEventApi.FINISHED_TASK_EVENT_PATH)
+interface FinishedTaskEventApiClient: FinishedTaskEventApi  {
 
-    val finishedTaskEvents: FinishedTaskEventApiClient
+    @GetExchange
+    override fun getEvents(@RequestParam timestamp: Instant, @ParameterObject paginationRequest: PaginationRequest): FinishedTaskEventsResponse
 }